Conner's updated pages
Rafael: I have made some final updates here before sharing with you!
Some notes…

1. I have included a “scrollspy” plugin which updates the navbar as you
scroll down the page (to continue this for further sections, simply be
sure that the id of the a tag corresponds with the href attribute in
the navbar link.

2. I have included a jQuery plugin at the bottom as well ;) for fun

3. I could not get the navbar links (including the button “Find Out
More” on the jumbotron) to properly line up for some reason—>perhaps
there is still a margins error? I am just too tired rn to figure it
out, sorry :(

4. The charity images at the bottom of the screen correspond with the
specific image files: (charity name)copy.(end) in the img folder…they
have been specifically cropped, so pls don’t delete them if you want to
avoid formatting issues, as they are almost exact copies of the
original.

5. On the “connors-pages.html” I have commented out an entire section
where I was trying to add a carousel for the “founding members” section
(it didn’t work out unfortunately)…if you still want to try that, feel
free; if not, it can be deleted.

6. That’s all I believe. Feel free to play around, but all that needs
to be done is some minor formatting (as well as the addition of the
remaining parts, ie. founding members, sponsors, donate)…and yeah! Pls
let me know how it works out, and if you could send me a final version
when its all done, I’d really appreciate it ;)